+++ adif/util_collect_fpd_v2.r
@@ -11,7 +11,9 @@
 # Output: Checkpoint CSVs after each phase for inspection
 ################################################################################
 
-#### LOAD LIBRARIES ####
+# Header & Config ----
+## Header ----
+#### * LOAD LIBRARIES ####
 library(googledrive)
 library(googlesheets4)
 library(dplyr)
@@ -22,7 +24,7 @@ library(janitor)
 
 cat ("\n-----------\nADIF first party data pipeline started at:", format(Sys.time(), "%Y-%m-%d %H:%M:%S"), "\n-----------\n")
 
-#### CONFIGURATION ####
+####* CONFIGURATION ####
 gdrive_folder_id <- "1EyN93JE7v4OXjMMQREVuZ4ZN7xEed5WB"
 pattern <- "De Beers | Partner Data"
 output_dir <- "/Users/eugenetsenter/Looker_clonedRepo/looker_personal/adif/data"